% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_chat.R
\name{addHistory}
\alias{addHistory}
\title{Add history for chating context}
\usage{
addHistory(history, role, item)
}
\arguments{
\item{history}{The history of chat}

\item{role}{The role of chat: "user" or "model"}

\item{item}{The item of chat: "prompt" or "output"}
}
\value{
The history of chat
}
\description{
Add history for chating context
}
