% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_audio.R
\name{gemini_audio.vertex}
\alias{gemini_audio.vertex}
\title{Analyze Audio using Gemini Vertex API}
\usage{
gemini_audio.vertex(
  audio = NULL,
  prompt = "Describe this audio",
  tokens = NULL
)
}
\arguments{
\item{audio}{Path to the audio file (character string). only supports "mp3".}

\item{prompt}{A prompt to guide the Gemini API's analysis (character string, defaults to "Describe this audio").}

\item{tokens}{A list containing the API URL and key from token.vertex() function.}
}
\value{
A character vector containing the Gemini API's description of the audio.
}
\description{
This function sends audio to the Gemini API and returns a text description.
}
