% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_image.R
\name{gemini_image}
\alias{gemini_image}
\title{Generate text from text and image with Gemini}
\usage{
gemini_image(image = NULL, prompt = "Explain this image")
}
\arguments{
\item{image}{The image to generate text}

\item{prompt}{The prompt to generate text, Default is "Explain this image"}
}
\value{
Generated text
}
\description{
Generate text from text and image with Gemini
}
\examples{
\dontrun{
library(gemini.R)
setAPI("YOUR_API_KEY")
gemini_image(image = system.file("docs/reference/figures/image.png", package = "gemini.R"))
}

}
\seealso{
https://ai.google.dev/docs/gemini_api_overview#text_image_input
}
