% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_searchR.R
\name{gemini_searchR}
\alias{gemini_searchR}
\title{Generate text with real-time information using Gemini (Retrieval)}
\usage{
gemini_searchR(
  prompt,
  model = "1.5-flash",
  temperature = 1,
  maxOutputTokens = 8192,
  topK = 40,
  topP = 0.95,
  seed = 1234
)
}
\arguments{
\item{prompt}{The prompt or question to ask}

\item{model}{The model to use. Options are "1.5-flash", "1.5-pro". Default is '1.5-flash'.
see https://ai.google.dev/gemini-api/docs/models/gemini}

\item{temperature}{The temperature to use. Default is 1 value should be between 0 and 2
see https://ai.google.dev/gemini-api/docs/models/generative-models#model-parameters}

\item{maxOutputTokens}{The maximum number of tokens to generate.
Default is 8192 and 100 tokens correspond to roughly 60-80 words.}

\item{topK}{The top-k value to use. Default is 40 value should be between 0 and 100
see https://ai.google.dev/gemini-api/docs/models/generative-models#model-parameters}

\item{topP}{The top-p value to use. Default is 0.95 value should be between 0 and 1
see https://ai.google.dev/gemini-api/docs/models/generative-models#model-parameters}

\item{seed}{The seed to use. Default is 1234 value should be integer
see https://ai.google.dev/gemini-api/docs/models/generative-models#model-parameters}
}
\value{
Generated text response from the Gemini model
}
\description{
Generate text responses with simplified access to Gemini models
}
\examples{
\dontrun{
library(gemini.R)
setAPI("YOUR_API_KEY")
gemini_searchR("Who won the latest F1 grand prix?")
}
}
\seealso{
https://ai.google.dev/docs/gemini_api_overview#text_input
}
