% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma.R
\name{gemma}
\alias{gemma}
\title{Generate text from text with Gemma}
\usage{
gemma(prompt, model = "gemma-3-1b-it", api_key = NULL, timeout = 60)
}
\arguments{
\item{prompt}{The prompt to generate text from}

\item{model}{The model to use. Default is 'gemma-3-1b-it'.
see https://ai.google.dev/gemma/docs/get_started#models-list}

\item{api_key}{Your API key. If NULL, uses GEMINI_API_KEY environment variable.}

\item{timeout}{Request timeout in seconds. Default is 60.}
}
\value{
Generated text
}
\description{
Generate text from text with Gemma with Gemini API
}
\examples{
\dontrun{
gemma("Roses are red...")
}
}
