% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tests.R
\name{gen_tests}
\alias{gen_tests}
\title{Generates unit test code for an R function.}
\usage{
gen_tests(prompt = NULL)
}
\arguments{
\item{prompt}{A character string specifying the prompt for the Gemini model.  If NULL, a default prompt is used.}
}
\value{
#' A character string containing the generated unit test code.
}
\description{
Generates unit test code for an R function.
}
