% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_body}
\alias{init_body}
\title{Initialize Gemini API request body}
\usage{
init_body()
}
\value{
A list representing an empty Gemini API request body.
}
\description{
Creates an empty body structure for Gemini API requests.
}
\keyword{internal}
