% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAPI.R
\name{setAPI}
\alias{setAPI}
\title{Set Gemini API Key}
\usage{
setAPI(api_key)
}
\arguments{
\item{api_key}{A character string containing your Gemini API key.}
}
\value{
No return value, called for side effects.
}
\description{
Sets the Gemini API key as an environment variable for use in API calls.
}
\note{
Please be aware you have to agree to the terms of service of the API provider.
Any app that uses the API key is subject to the terms of service.
Also, please be aware that the API key is a sensitive information.
}
\examples{
\dontrun{
setAPI("YOUR_API_KEY")
}
}
\seealso{
https://makersuite.google.com/app/apikey
}
