% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini}
\alias{gemini}
\title{Generate text from text with Gemini}
\usage{
gemini(prompt)
}
\arguments{
\item{prompt}{The prompt to generate text from}
}
\value{
Generated text
}
\description{
Generate text from text with Gemini
}
\examples{
library(gemini.R)
setAPI("YOUR_API_KEY")
gemini("Explain dplyr's mutate function")

}
\seealso{
https://ai.google.dev/docs/gemini_api_overview#text_input
}
