% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gemini.vertex}
\alias{gemini.vertex}
\title{Generate text from text with Gemini Vertex API}
\usage{
gemini.vertex(prompt = NULL, tokens = NULL)
}
\arguments{
\item{prompt}{A character string containing the prompt for the Gemini model.}

\item{tokens}{A list containing the API URL and key from token.vertex() function.}
}
\value{
#' A character string containing the generated text.
}
\description{
Generate text from text with Gemini Vertex API
}
\examples{
\dontrun{
# token should be created before this. using the token.vertex() function
prompt <- "What is sachins Jersey number?"
gemini.vertex(prompt, tokens)
}

}
\seealso{
https://ai.google.dev/docs/gemini_api_overview#text_input
}
