% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_chat.R
\name{gemini_chat}
\alias{gemini_chat}
\title{Multi-turn conversations (chat)}
\usage{
gemini_chat(prompt, history = list())
}
\arguments{
\item{prompt}{The prompt to generate text from}
}
\value{
Generated text
}
\description{
Generate text from text with Gemini
}
\examples{
library(gemini.R)
setAPI("YOUR_API_KEY")

chats <- gemini_chat("Pretend you're a snowman and stay in character for each")
print(chats$outputs)

chats <- gemini_chat("What's your favorite season of the year?", chats$history)
print(chats$outputs)

chats <- gemini_chat("How do you think about summer?", chats$history)
print(chats$outputs)

}
\seealso{
https://ai.google.dev/docs/gemini_api_overview#chat
}
