% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_image.R
\name{gemini_image}
\alias{gemini_image}
\title{Generate text from text and image with Gemini}
\usage{
gemini_image(prompt, image)
}
\arguments{
\item{prompt}{The prompt to generate text}

\item{image}{The image to generate text}
}
\value{
Generated text
}
\description{
Generate text from text and image with Gemini
}
\examples{
library(gemini.R)
setAPI("YOUR_API_KEY")
gemini_image("Explain this image", "https://upload.wikimedia.org/wikipedia/commons/thumb/4/44/Sidney_Hall_-_Urania\%27s_Mirror_-_Gemini.jpg/560px-Sidney_Hall_-_Urania\%27s_Mirror_-_Gemini.jpg")

}
\seealso{
https://ai.google.dev/docs/gemini_api_overview#text_image_input
}
