% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_image.R
\name{gemini_image.vertex}
\alias{gemini_image.vertex}
\title{Generate text from text and image with Gemini Vertex API}
\usage{
gemini_image.vertex(
  image = NULL,
  prompt = "Explain this image",
  type = "png",
  tokens = NULL
)
}
\arguments{
\item{image}{The image to generate text}

\item{prompt}{A character string specifying the prompt to use with the image. Defaults to "Explain this image".  Currently ignored.}

\item{type}{A character string specifying the image type ("png", "jpeg", "webp", "heic", "heif"). Defaults to "png".}

\item{tokens}{A list containing the API URL and key from token.vertex() function.}
}
\value{
A character string containing Gemini's description of the image.
}
\description{
Generate text from text and image with Gemini Vertex API
}
