% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_docs.R
\name{gen_docs}
\alias{gen_docs}
\title{Generate Roxygen Documentation}
\usage{
gen_docs(prompt = NULL)
}
\arguments{
\item{prompt}{A character string specifying additional instructions for the LLM.  Defaults to a prompt requesting Roxygen2 documentation without the original code.}
}
\value{
#' A character string containing the generated Roxygen2 documentation.  The documentation will start with \verb{#'}.
}
\description{
Generates Roxygen2 documentation for an R function based on the currently selected code.
}
