% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAPI.R
\name{setAPI}
\alias{setAPI}
\title{Set API key}
\usage{
setAPI(api_key)
}
\arguments{
\item{api_key}{The API key to set}
}
\description{
Set API key as an environment variable
}
\examples{
\dontrun{
setAPI("my_api_key")
}
}
\seealso{
https://makersuite.google.com/app/apikey
}
\keyword{internal}
